+++ modules/colors.py
@@ -1,0 +1,5 @@
+class Color:
+    WHITE = (255, 255, 255)
+    BLACK = (0, 0, 0)
+    RED = (204, 49, 61)
+    BLUE = (64, 142, 198)
