+++ modules/config.py
@@ -1,0 +1,3 @@
+class Config:
+    WINDOW_WIDTH = 1920
+    WINDOW_HEIGHT = 1080
